# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Falzoni PHP Laravel",
        "description": "Api de serviços com Laravel PHP",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customer/get/{id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Listar cliente por Id",
                "description": "Retorna o objeto do cliente",
                "operationId": "3d5760bcdaf41adaed1c79513098f140",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitação efetuada com sucesso"
                    }
                }
            }
        },
        "/api/customer/create": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Adicionar cliente",
                "description": "Cria um novo registro de cliente",
                "operationId": "a9141f7af7e66c903e7253b2c71f4266",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "document": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "name": "Test Customer",
                                    "document": "123.456.789-00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/customer/update": {
            "put": {
                "tags": [
                    "Customer"
                ],
                "summary": "Atualizar cliente",
                "description": "Atualiza um registro de cliente",
                "operationId": "cefeb54afd68315899fbd1d742c59e73",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "document": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "name": "Test Customer",
                                    "document": "123.456.789-00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/customer/delete/{id}": {
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "Remover cliente por Id",
                "description": "Remove o cliente",
                "operationId": "b9badb2c165c775507b2fb29cb8776c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
...
        "/api/product/get/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Listar produto por Id",
                "description": "Retorna o objeto do produto",
                "operationId": "bfa77bf640e9baf3b428ea6473ff98b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitação efetuada com sucesso"
                    }
                }
            }
        },
        "/api/product/create": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Adicionar produto",
                "description": "Cria um novo registro de produto",
                "operationId": "332e45b9c2a63faa0747f6cb178b9f93",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "document": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "name": "Test Product",
                                    "document": "123.456.789-00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/product/update": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Atualizar produto",
                "description": "Atualiza um registro de produto",
                "operationId": "7832bd7edc0a3d4bb6b557f467ac7d3d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "document": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "name": "Test Product",
                                    "document": "123.456.789-00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/product/delete/{id}": {
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Remover produto por Id",
                "description": "Remove o produto",
                "operationId": "ad920499e86c324625ef4380d13d9629",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "Token de autorização (JWT) a ser inserido",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
...
}